requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlmodel>=0.0.8
alembic>=1.7.7
python-dotenv>=0.19.0
pydantic>=1.8.0,<2.0.0
typing-extensions>=4.5.0
passlib>=1.7.4
bcrypt>=3.2.0
email-validator>=1.1.3
psycopg2-binary>=2.9.1